requirements.txt
Flask
flask-cors
gunicorn
yt-dlp
moviepy
scenedetect[opencv]
google-generativeai
webvtt-py